+++ src/config.js
@@ -10,11 +10,12 @@
  * governing permissions and limitations under the License.
  */
 
-import { errorWithResponse } from './util.js';
+import { errorWithResponse } from './utils/http.js';
 
 /**
- * @param {string[]} patterns
- * @param {string} path
+ * This function finds ordered matches between a list of patterns and a given path.
+ * @param {string[]} patterns - An array of pattern strings to match against.
+ * @param {string} path - The path string to match patterns against.
  */
 function findOrderedMatches(patterns, path) {
   return patterns
@@ -28,9 +29,10 @@ function findOrderedMatches(patterns, path) {
 }
 
 /**
- * @param {string} pattern
- * @param {string} path
- * @returns {Record<string, string>}
+ * This function extracts path parameters from a pattern and a path.
+ * @param {string} pattern - The pattern string.
+ * @param {string} path - The path string.
+ * @returns {Record<string, string>} - The path parameters.
  */
 function extractPathParams(pattern, path) {
   // create a RegExp with named groups from the string contained in '{{}}'
@@ -40,10 +42,10 @@ function extractPathParams(pattern, path) {
 }
 
 /**
- * @param {Context} ctx
- * @param {string} tenant
- * @param {Partial<Config>} [overrides={}]
- * @returns {Promise<Config|null>}
+ * This function resolves the configuration for a given context and overrides.
+ * @param {Context} ctx - The context object.
+ * @param {Partial<Config>} [overrides={}] - The overrides object.
+ * @returns {Promise<Config|null>} - A promise that resolves to the configuration.
  */
 export async function resolveConfig(ctx, overrides = {}) {
   const [_, org, site, route] = ctx.url.pathname.split('/');
@@ -58,6 +60,10 @@ export async function resolveConfig(ctx, overrides = {}) {
   }
 
   const siteKey = `${org}--${site}`;
+
+  /**
+   * @type {Config}
+   */
   const confMap = await ctx.env.CONFIGS.get(siteKey, 'json');
   if (!confMap) {
     return null;
@@ -99,30 +105,5 @@ export async function resolveConfig(ctx, overrides = {}) {
     route,
     ...overrides,
   };
-
-  // If the route is catalog, get the environment from the path segment
-  if (route === 'catalog') {
-    const pathSegments = ctx.url.pathname.split('/');
-    const catalogIndex = pathSegments.indexOf('catalog');
-
-    // Ensure that there are exactly 4 segments after 'catalog' (env, store, storeView, product)
-    if (catalogIndex !== -1 && pathSegments.length >= catalogIndex + 4) {
-      resolved.env = pathSegments[catalogIndex + 1];
-      resolved.storeCode = pathSegments[catalogIndex + 2];
-      resolved.storeViewCode = pathSegments[catalogIndex + 3];
-      resolved.subRoute = pathSegments[catalogIndex + 4];
-      resolved.sku = pathSegments[catalogIndex + 5];
-    } else {
-      throw new Error('Invalid URL structure: Missing required segments after "catalog". Expected format: /catalog/{env}/{store}/{storeView}/{product}[/{sku}]');
-    }
-  }
-
-  // ensure validity
-  // TODO: make this more robust
-  if (!resolved.pageType && route !== 'catalog') {
-    ctx.log.warn('invalid config for tenant site (missing pageType)', siteKey);
-    return null;
-  }
-
   return resolved;
 }
